backend/requirements.txt
fastapi
uvicorn[standard]
requests
pillow
python-dotenv
anthropic
python-multipart
aiofiles
